refactor(collector-core): split rpc_services.rs into rpc_services/ submodules
rpc_services.rs (883 lines) split into a module directory:
- rpc_services/config.rs    (24)  — RpcServiceConfig + Default
- rpc_services/providers.rs (277) — the three Collector*Provider impls
- rpc_services/manager.rs   (590) — CollectorRpcServiceManager + impl + test
- rpc_services/mod.rs        (24) — module doc + re-exports (public API preserved)

Pure mechanical extraction; zero logic/API changes. 146 lib tests + 21 doctests
pass, clippy -D warnings clean.
